dashboardImplementation: architectural reorganization
