+++ css/components/dashboard/FileItem.css
@@ -1,0 +1,4 @@
+.file-item {
+  text-align : left;
+  justify-content: space-evenly;
+}
