+++ css/components/dashboard/GradeWidget.css
@@ -1,0 +1,36 @@
+.rounded-rectangle {
+  border-radius: 15px;
+  background-color: white;
+  height: 210px;
+  width: 500px;
+  margin-left: 80px;
+  margin-top: 40px;
+}
+
+.grade {
+  font-size: 30px;
+  font-weight: bold;
+  color: #515151 !important;
+}
+
+.grade-pos {
+  top: 35px;
+  left: 305px;
+}
+
+.grade-text {
+  font-size: 14px;
+}
+
+.grade-text-pos {
+  position: absolute;
+  top: 35% !important;
+  left: 70px;
+  width: 160px !important;
+  text-align: left;
+}
+
+.divider-pos {
+  top: 0px !important;
+  position: absolute;
+}
