+++ css/components/dashboard/GraphWidget.css
@@ -1,0 +1,24 @@
+.rounded-rectangle {
+  border-radius: 15px;
+  background-color: white;
+  height: 210px;
+  width: 500px;
+  margin-left: 80px;
+}
+
+.centered-container-graph {
+  margin: 10px !important;
+}
+
+.widget-title {
+  position: absolute;
+  top: 5px;
+  left: 8px;
+  font-size: 14px;
+  color: #515151;
+}
+
+.centered-text {
+  text-align: center;
+  margin-top: 30px;
+}
