--- src/components/css/AddButton.css
@@ -1,12 +1,0 @@
-.add-button {
-  position: fixed;
-  right: 20px;
-  bottom: 20px;
-  border-radius: 50%;
-  width: 30px;
-  height: 30px;
-  min-width: 30px;
-  font-size: larger;
-  cursor: pointer;
-  border: none;
-}
