+++ src/components/dashboard/widgets/OverView/OverViewWidget.vue
@@ -71,39 +71,4 @@ export default {
 };
 </script>
 
-<style scoped>
-.rounded-rectangle {
-  border-radius: 15px;
-  background-color: white;
-  height: 460px;
-  width: 770px;
-  margin-right: 40px;
-}
-
-.custom-tabs {
-  display: relative;
-}
-
-.active-tab {
-  background-color: white !important;
-}
-
-.custom-tabs .v-tab {
-  flex: 1;
-  background-color: #E8E8E8;
-}
-
-.filter-button {
-  position: fixed;
-  left: 20px;
-  bottom: 20px;
-  border-radius: 50%;
-  width: 30px;
-  height: 30px;
-  min-width: 30px;
-  font-size: larger;
-  cursor: pointer;
-  border: none;
-}
-
-</style>
+<style src="@/../css/components/dashboard/OverViewWidget.css" scoped></style>
