--- src/components/widgets/css/FileList.css
@@ -1,99 +1,0 @@
-.file_list {
-  height: auto;
-  border-radius: 8px;
-  background: #ffffff;
-  box-shadow:
-    9px 9px 18px #d9d9d9,
-    -9px -9px 18px #ffffff;
-  overflow-y: auto;
-  padding: 0;
-}
-
-.header {
-  display: block;
-  background: #ffee00;
-  padding: 15px;
-}
-
-.header h4 {
-  margin: 0;
-  padding: 0;
-}
-
-.file_item {
-  padding: 0;
-}
-
-.file_list .v-list-item:nth-child(odd) {
-  background-color: #fff56a;
-}
-
-.file-list li {
-  position: relative;
-}
-
-.file-item {
-  display: flex;
-  padding: 10px 0;
-  align-items: center;
-  flex-direction: column;
-}
-
-.file-name {
-  display: flex;
-  justify-content: space-evenly;
-  width: 100%;
-}
-
-.redirect-icon {
-  width: 16px;
-  height: 16px;
-  margin-left: 5px;
-  display: none;
-}
-
-.file-name.file-hovered + .redirect-icon {
-  display: inline-block;
-}
-
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
-}
-
-.widget-container {
-  margin-left: 300px;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.file-info {
-  margin-top: 20px;
-}
-
-.full-height {
-  height: 100%;
-}
-
-.edit-icon {
-  width: auto;
-  height: 17px;
-  cursor: pointer;
-  right: 20px;
-  position: fixed;
-}
-
-.emission_size {
-  font-size: xx-large;
-  font-weight: bolder;
-}
-
-.evolution_option {
-  display: grid;
-  text-align: center;
-  font-size: large;
-}
