--- src/components/widgets/exampleData/co2.json
@@ -1,5 +1,0 @@
-{
-    "emissions": "502 tCO2e",
-    "trend": "+ 12 %",
-    "showEvolution": false
-}
