+++ src/pages/NavBar.vue
@@ -1,7 +1,7 @@
 <template>
   <v-app-bar class="px-3" flat density="compact">
     <v-app-bar-icon>
-      <img src="../assets/tidybeelogo.png" />
+      <img src="@/assets/tidybeelogo.png" />
     </v-app-bar-icon>
     <v-spacer />
     <v-tabs centered color="grey-darken-2">
@@ -42,7 +42,7 @@ export default {
   data() {
     return {
       availableLocales: [""],
-      links: ["Home", "Configuration", "About"],
+      links: ["Home", "Settings", "About"],
     };
   },
   mounted() {
